// lib/api/generated/models/ConversationEntity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lumora API
 * Lumora backend REST API.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MessageEntity } from './MessageEntity';
import {
    MessageEntityFromJSON,
    MessageEntityFromJSONTyped,
    MessageEntityToJSON,
    MessageEntityToJSONTyped,
} from './MessageEntity';

/**
 * 
 * @export
 * @interface ConversationEntity
 */
export interface ConversationEntity {
    /**
     * 
     * @type {string}
     * @memberof ConversationEntity
     */
    created_at?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationEntity
     */
    id?: string;
    /**
     * 
     * @type {Array<MessageEntity>}
     * @memberof ConversationEntity
     */
    messages?: Array<MessageEntity>;
    /**
     * 
     * @type {string}
     * @memberof ConversationEntity
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationEntity
     */
    updated_at?: string;
    /**
     * 
     * @type {string}
     * @memberof ConversationEntity
     */
    user_id?: string;
}

/**
 * Check if a given object implements the ConversationEntity interface.
 */
export function instanceOfConversationEntity(value: object): value is ConversationEntity {
    return true;
}

export function ConversationEntityFromJSON(json: any): ConversationEntity {
    return ConversationEntityFromJSONTyped(json, false);
}

export function ConversationEntityFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConversationEntity {
    if (json == null) {
        return json;
    }
    return {
        
        'created_at': json['created_at'] == null ? undefined : json['created_at'],
        'id': json['id'] == null ? undefined : json['id'],
        'messages': json['messages'] == null ? undefined : ((json['messages'] as Array<any>).map(MessageEntityFromJSON)),
        'title': json['title'] == null ? undefined : json['title'],
        'updated_at': json['updated_at'] == null ? undefined : json['updated_at'],
        'user_id': json['user_id'] == null ? undefined : json['user_id'],
    };
}

export function ConversationEntityToJSON(json: any): ConversationEntity {
    return ConversationEntityToJSONTyped(json, false);
}

export function ConversationEntityToJSONTyped(value?: ConversationEntity | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['created_at'],
        'id': value['id'],
        'messages': value['messages'] == null ? undefined : ((value['messages'] as Array<any>).map(MessageEntityToJSON)),
        'title': value['title'],
        'updated_at': value['updated_at'],
        'user_id': value['user_id'],
    };
}