// lib/api/generated/models/MessageEntity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lumora API
 * Lumora backend REST API.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Role } from './Role';
import {
    RoleFromJSON,
    RoleFromJSONTyped,
    RoleToJSON,
    RoleToJSONTyped,
} from './Role';
import type { MessageKind } from './MessageKind';
import {
    MessageKindFromJSON,
    MessageKindFromJSONTyped,
    MessageKindToJSON,
    MessageKindToJSONTyped,
} from './MessageKind';

/**
 * 
 * @export
 * @interface MessageEntity
 */
export interface MessageEntity {
    /**
     * 
     * @type {string}
     * @memberof MessageEntity
     */
    ai_request_id?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageEntity
     */
    content?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageEntity
     */
    conversation_id?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageEntity
     */
    created_at?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageEntity
     */
    id?: string;
    /**
     * 
     * @type {MessageKind}
     * @memberof MessageEntity
     */
    kind?: MessageKind;
    /**
     * 
     * @type {Array<number>}
     * @memberof MessageEntity
     */
    payload?: Array<number>;
    /**
     * 
     * @type {Role}
     * @memberof MessageEntity
     */
    role?: Role;
    /**
     * 
     * @type {string}
     * @memberof MessageEntity
     */
    updated_at?: string;
}



/**
 * Check if a given object implements the MessageEntity interface.
 */
export function instanceOfMessageEntity(value: object): value is MessageEntity {
    return true;
}

export function MessageEntityFromJSON(json: any): MessageEntity {
    return MessageEntityFromJSONTyped(json, false);
}

export function MessageEntityFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessageEntity {
    if (json == null) {
        return json;
    }
    return {
        
        'ai_request_id': json['ai_request_id'] == null ? undefined : json['ai_request_id'],
        'content': json['content'] == null ? undefined : json['content'],
        'conversation_id': json['conversation_id'] == null ? undefined : json['conversation_id'],
        'created_at': json['created_at'] == null ? undefined : json['created_at'],
        'id': json['id'] == null ? undefined : json['id'],
        'kind': json['kind'] == null ? undefined : MessageKindFromJSON(json['kind']),
        'payload': json['payload'] == null ? undefined : json['payload'],
        'role': json['role'] == null ? undefined : RoleFromJSON(json['role']),
        'updated_at': json['updated_at'] == null ? undefined : json['updated_at'],
    };
}

export function MessageEntityToJSON(json: any): MessageEntity {
    return MessageEntityToJSONTyped(json, false);
}

export function MessageEntityToJSONTyped(value?: MessageEntity | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'ai_request_id': value['ai_request_id'],
        'content': value['content'],
        'conversation_id': value['conversation_id'],
        'created_at': value['created_at'],
        'id': value['id'],
        'kind': MessageKindToJSON(value['kind']),
        'payload': value['payload'],
        'role': RoleToJSON(value['role']),
        'updated_at': value['updated_at'],
    };
}